%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  NativeCode.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{NativeCode}
\docType{class}
\alias{NativeCode}


\title{Static class to compile and loaded native code}

\description{
 Package:  R.native \cr
\bold{Class NativeCode}\cr

\code{\link[R.oo]{Object}}\cr
\code{~~|}\cr
\code{~~+--}\code{NativeCode}\cr

\bold{Directly known subclasses:}\cr
\cr

public static class \bold{NativeCode}\cr
extends \link[R.oo]{Object}\cr


}

\section{Fields and Methods}{
 \bold{Methods:}\cr
\tabular{rll}{
 \tab \code{\link[R.native:compile.NativeCode]{compile}} \tab Compiles native code.\cr
 \tab \code{\link[R.native:dynLoad.NativeCode]{dynLoad}} \tab Static method to load a dynamic library.\cr
 \tab \code{\link[R.native:dynReload.NativeCode]{dynReload}} \tab Static method to reload a loaded dynamic library.\cr
 \tab \code{\link[R.native:dynUnload.NativeCode]{dynUnload}} \tab Static method to unload a dynamic library.\cr
 \tab \code{\link[R.native:isLoaded.NativeCode]{isLoaded}} \tab Static method to check if a library is loaded.\cr
 \tab \code{\link[R.native:moveLibs.NativeCode]{moveLibs}} \tab Moves library files.\cr
}


\bold{Methods inherited from Object}:\cr
$, $<-, [[, [[<-, as.character, attach, attachLocally, clearCache, clone, detach, equals, extend, finalize, gc, getEnvironment, getFields, getInstanciationTime, getStaticInstance, hasField, hashCode, ll, load, objectSize, print, save


}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{classes}
\keyword{programming}
