%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{R.native-package}
\alias{R.native-package}
\alias{R.native}
\docType{package}

\title{Package R.native}


\description{
  This packages provides a utility class NativeCode that provides static methods to compile native code into a dynamic library, and methods to load, reload and unload such.

  This package should be considered to be in "alpha" version.
}

\section{Installation}{
  To install this package, see \url{http://www.braju.com/R/}.
  Required packages are installed in the same way.
}

\section{Dependancies and other requirements}{
  This package depends on the \pkg{R.oo} and \pkg{R.utils} packages.
  It is uses native code and is therefore not necessarily working
  on all platforms.
}

\section{To get started}{
  To get started, see:
  \enumerate{
    \item \code{\link{sourceC}}() - To compile C code and load it into R.
    \item \code{\link{NativeCode}} - A utility class for handling native code.
  }
}

\section{How to cite this package}{
  Not available.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}
\keyword{package}
